# Propagate iteration info to optimizer